HW_4 edited exceptions (created NotFoundException)

+++ src/main/java/com/epam/springshop/service/impl/UserServiceImpl.java
@@ -82,6 +82,9 @@ public class UserServiceImpl implements UserService {
     @Override
     public void deleteUser(Long obj) {
         log.info(String.format("%s : method ==> deleteUser(%s)", this.getClass().getName(), obj));
+        if (userRepoImpl.read(obj) == null) {
+            throw new UserNotFoundException();
+        }
         userRepoImpl.delete(obj);
     }
 }
